refactor: interop/lyric-fetch/lyric-scroll
