+++ src/sync/lyric/mod.rs
@@ -1,0 +1,2 @@
+pub mod fetch;
+pub mod scroll;
