+++ src/sync/mod.rs
@@ -6,21 +6,13 @@ use mpris::{Player, PlayerFinder};
 use crate::lyric::LyricOwned;
 
 mod utils;
-mod lyric;
-mod player;
+pub mod lyric;
 mod cache;
 mod search_window;
+mod interop;
 
-pub use lyric::register_lyric_display;
-pub use player::register_mpris_sync;
-pub use player::register_sigusr1_disconnect;
-pub use player::register_action_disconnect;
-pub use player::register_action_connect;
-pub use player::register_action_search_lyric;
-pub use player::register_action_remove_lyric;
-pub use player::register_action_reload_lyric;
-
-pub(crate) use player::list_avaliable_players;
+pub use lyric::scroll::register_lyric_display;
+pub(crate) use interop::list_avaliable_players;
 
 /// A struct from metadata in mpris::TrackID to avoid track_id and title unwrapping
 #[derive(Clone, Debug)]
@@ -38,3 +30,10 @@ thread_local! {
     /// including: track_id, paused, cache_path
     static TRACK_PLAYING_STATE: RefCell<(Option<TrackMeta>, bool, Option<PathBuf>)> = RefCell::new((None, false, None));
 }
+
+pub use interop::acts::{
+    register_action_connect, register_action_disconnect, register_action_search_lyric,
+    register_action_remove_lyric, register_action_reload_lyric, register_sigusr1_disconnect,
+};
+
+pub use interop::register_mpris_sync;
